requirements.txt
django~=4.2.11
pillow~=10.3.0
qrcode~=7.4.2
django-ckeditor~=6.7.1